+++ tests/testthat/test-verbs.R
@@ -178,6 +178,16 @@ test_that("set_edges replaces any existing edges for pairs", {
   )
 })
 
+test_that("set_edges removes both directions in simple graphs", {
+  cg <- caugi(A %-->% B, class = "DAG")
+  cg1 <- set_edges(cg, B %-->% A)
+
+  expect_equal(
+    cg1@edges,
+    data.table::data.table(from = "B", edge = "-->", to = "A")
+  )
+})
+
 test_that("set_edges errors whwn both vector and expr paths are given", {
   # Use simple=FALSE to allow parallel edges for this test
   cg <- caugi(simple = FALSE, class = "UNKNOWN")
@@ -541,11 +551,29 @@ test_that("AUTO class doesn't trigger after init", {
   expect_equal(cg@graph_class, "DAG")
 })
 
-test_that("node order does not matter when removing edges", {
+test_that("node order does not matter when setting edges", {
   cg <- caugi(A %---% B, class = "PDAG")
   cg_new <- set_edges(cg, B %-->% A)
 
   cg_new
 
   expect_equal(edges(cg_new)$edge, "-->")
+})
+
+test_that("set_edges should not replace other edges when non-symmetric edge added", {
+  cg <- caugi(A %-->% B, simple = FALSE, class = "UNKNOWN")
+  cg_new <- set_edges(cg, B %-->% A)
+
+  cg_correct <- caugi(A %-->% B, B %-->% A, simple = FALSE, class = "UNKNOWN")
+
+  expect_equivalent(cg_new, cg_correct)
+})
+
+test_that("set_edges preserves opposite-direction edges in non-simple graphs", {
+  cg <- caugi(A %-->% B, B %<->% A, simple = FALSE, class = "UNKNOWN")
+  cg_new <- set_edges(cg, A %o->% B)
+
+  cg_correct <- caugi(A %o->% B, B %<->% A, simple = FALSE, class = "UNKNOWN")
+
+  expect_equivalent(cg_new, cg_correct)
 })